chatbote/requirements.txt
streamlit
langchain
openai
python-dotenv
faiss-cpu
PyPDF2
tiktoken